+++ cpp/subprojects/boosting/include/boosting/output/predictor_regression_label_wise.hpp
@@ -34,6 +34,11 @@ namespace boosting {
             std::unique_ptr<IRegressionPredictorFactory> createRegressionPredictorFactory(
                 const IFeatureMatrix& featureMatrix, uint32 numLabels) const override;
 
+            /**
+             * @see `IPredictorConfig::isLabelVectorSetNeeded`
+             */
+            bool isLabelVectorSetNeeded() const override;
+
     };
 
 }