+++ cpp/subprojects/boosting/src/boosting/rule_evaluation/rule_evaluation_label_wise_common.hpp
@@ -63,23 +63,4 @@ namespace boosting {
         return qualityScore + l1RegularizationTerm + l2RegularizationTerm;
     }
 
-    /**
-     * Calculates and returns a quality score that assesses the quality of the optimal prediction for a single label.
-     *
-     * @param gradient                  The gradient that corresponds to the label
-     * @param hessian                   The Hessian that corresponds to the label
-     * @param l1RegularizationWeight    The weight of the L1 regularization
-     * @param l2RegularizationWeight    The weight of the L2 regularization
-     * @return                          The quality score that has been calculated
-     */
-    static inline constexpr float64 calculateLabelWiseQualityScore(float64 gradient, float64 hessian,
-                                                                   float64 l1RegularizationWeight,
-                                                                   float64 l2RegularizationWeight) {
-        float64 l1Weight = getL1RegularizationWeight(gradient, l1RegularizationWeight);
-        float64 l1Term = l1Weight != 0
-                            ? ((2 * gradient * l1Weight) - (3 * l1RegularizationWeight * l1RegularizationWeight))
-                            : (-gradient * l1RegularizationWeight);
-        return divideOrZero(-0.5 * (gradient * gradient + l1Term), hessian + l2RegularizationWeight);
-    }
-
 }