+++ cpp/subprojects/common/include/common/output/predictor_regression.hpp
@@ -49,7 +49,7 @@ class IRegressionPredictorFactory {
  * Defines an interface for all classes that allow to configure a predictor that predicts label-wise regression scores
  * for given query examples.
  */
-class IRegressionPredictorConfig {
+class IRegressionPredictorConfig : public IPredictorConfig {
 
     public: