+++ python/subprojects/testbed/mlrl/testbed/args.py
@@ -41,7 +41,11 @@ PARAM_PARAMETER_DIR = '--parameter-dir'
 
 PARAM_OUTPUT_DIR = '--output-dir'
 
+PARAM_PRINT_PREDICTIONS = '--print-predictions'
+
 PARAM_STORE_PREDICTIONS = '--store-predictions'
+
+PARAM_PREDICT_PROBABILITIES = '--predict-probabilities'
 
 PARAM_PRINT_DATA_CHARACTERISTICS = '--print-data-characteristics'
 
@@ -185,10 +189,16 @@ def add_learner_arguments(parser: ArgumentParser):
                              + 'used by the algorithm, are located.')
     parser.add_argument(PARAM_OUTPUT_DIR, type=str,
                         help='The path of the directory where experimental results should be saved.')
+    parser.add_argument(PARAM_PRINT_PREDICTIONS, type=boolean_string, default=False,
+                        help='Whether the predictions for individual examples and labels should be printed on the ' +
+                             'console or not. Must be one of ' + format_enum_values(BooleanOption) + '.')
     parser.add_argument(PARAM_STORE_PREDICTIONS, type=boolean_string, default=False,
                         help='Whether the predictions for individual examples and labels should be written into output '
                              + 'files or not. Must be one of ' + format_enum_values(BooleanOption) + '. Does only have '
                              + 'an effect, if the parameter ' + PARAM_OUTPUT_DIR + ' is specified.')
+    parser.add_argument(PARAM_PREDICT_PROBABILITIES, type=boolean_string, default=False,
+                        help='Whether probabilities should be predicted rather than binary labels or not. Must be one '
+                            + 'of ' + format_enum_values(BooleanOption) + '.')
 
 
 def add_rule_learner_arguments(parser: ArgumentParser):